Raw import of GM7 source
Raw import of the source from the GM7 version.
To import a GM7 game into GameMaker, I first imported it into GameMaker Studio 1.4. Then I imported that conversion omtp modern GameMaker.
I am not including the GameMaker Studio 1.4 conversion in this repository, as it is only used as a stop-gap, and won't be needed after this.